Ejercicio extra 3 guia 12

+++ src/guia10/ejercicio3/services/AlumnoService.java
@@ -51,7 +51,7 @@ public class AlumnoService {
             }
         }
 
-        promedio = (double) (notaTotal / 3);
+        promedio = notaTotal / 3;
 
         if (getAlumno != null) {
             System.out.println("La nota promedio final de " + getAlumno.getNombre() + " es: " + promedio);
